refactor: create quic socket explicitly
