+++ perf/src/bin/tcp.rs
@@ -105,7 +105,7 @@ async fn test_shadowquic() {
         zero_rtt: true,
         over_stream: true,
         ..Default::default()
-    });
+    }).unwrap();
 
     let client = Manager {
         inbound: Box::new(socks_server),
