+++ perf/src/bin/udp.rs
@@ -125,7 +125,7 @@ async fn test_shadowquic() {
         zero_rtt: true,
         over_stream: false,
         ..Default::default()
-    });
+    }).unwrap();
 
     let client = Manager {
         inbound: Box::new(socks_server),
