+++ shadowquic/examples/client_server.rs
@@ -53,7 +53,8 @@ async fn test_shadowquic() {
         zero_rtt: true,
         over_stream: false,
         ..Default::default()
-    });
+    })
+    .unwrap();
 
     let client = Manager {
         inbound: Box::new(socks_server),
