+++ shadowquic/tests/dns_udp.rs
@@ -257,7 +257,8 @@ async fn shadowquic_client_server(over_stream: bool, port: u16) {
         zero_rtt: true,
         over_stream,
         ..Default::default()
-    });
+    })
+    .unwrap();
 
     let client = Manager {
         inbound: Box::new(socks_server),
