+++ shadowquic/tests/tcp_echo.rs
@@ -110,7 +110,8 @@ async fn test_shadowquic() {
         zero_rtt: true,
         over_stream: true,
         ..Default::default()
-    });
+    })
+    .unwrap();
 
     let client = Manager {
         inbound: Box::new(socks_server),
