+++ shadowquic/tests/udp_overdatagram_echo.rs
@@ -129,7 +129,8 @@ async fn test_shadowquic() {
         zero_rtt: true,
         over_stream: false,
         ..Default::default()
-    });
+    })
+    .unwrap();
 
     let client = Manager {
         inbound: Box::new(socks_server),
