+++ shadowquic/tests/udp_overstream_echo.rs
@@ -113,7 +113,8 @@ async fn test_shadowquic() {
         zero_rtt: true,
         over_stream: true,
         ..Default::default()
-    });
+    })
+    .unwrap();
 
     let client = Manager {
         inbound: Box::new(socks_server),
